// resources/gagot-api/src/enums/model.enum.ts
export enum ModelEnum {
  Articles = 'articles',
  Categories = 'categories',
  Users = 'users',
  Visits = 'visits',
  Complains = 'complains',
  Reviews = 'reviews',
  Properties = 'properties',
  Roles = 'roles',
  Rooms = 'rooms',
  Recent = 'recent',
  ForgotPassword = 'forgotPassword',
  ResetPassword = 'resetPassword',
  Files = 'files',
  Cache = 'cache',
}